# openapi.json
{
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "properties": {
          "login": {
            "description": "Email or username.",
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "login",
          "password"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 `application/problem+json` error body shape produced by [`ApiError`]. Declared\npurely for `OpenAPI` documentation — [`ApiError::into_response`] builds the JSON by hand so\nruntime callers never construct this type.",
        "example": {
          "detail": "resource not found",
          "status": 404,
          "title": "not_found",
          "type": "about:blank#not_found"
        },
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "format": "int64",
            "type": "integer"
          },
          "token_type": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "TankoVault contributors"
    },
    "description": "Axum HTTP edge: public read/write + admin, auth, resolved links, SSE.",
    "license": {
      "name": ""
    },
    "title": "tankovault-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "description": "Authenticates by email or username + password. Issues an access token and a rotating\nrefresh-token cookie.\n\nEvery outcome — success, unknown identifier, bad password, unverified address — is\naudited. An authentication log that only records successes cannot answer the question\nanyone actually asks it after an incident.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Authenticated; access token issued"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid email/username or password"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Email address not yet confirmed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too many attempts; retry later"
          }
        },
        "summary": "Log in",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/password/forgot": {
      "post": {
        "description": "Always responds `202 Accepted`, whether or not the address is registered, so the\nendpoint can't be used to probe which emails have accounts. When the address does exist\nand email is configured, a single-use, time-limited reset link is sent.",
        "operationId": "forgot_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "If the address is registered, a reset email has been sent"
          }
        },
        "summary": "Request a password-reset email",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "description": "Consumes a valid, unexpired, unused reset token, sets the new password, and revokes all\nof the user's active sessions so any stolen refresh token dies with the old credential.",
        "operationId": "reset_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed; existing sessions revoked"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid/expired token or weak password"
          }
        },
        "summary": "Reset a password with a token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Reads the `refresh_token` `HttpOnly` cookie, rotates it, and issues a fresh access token.\nPresenting a token that was already rotated (reuse) revokes the whole token family.",
        "operationId": "refresh",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Refreshed; new access token issued"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, expired, or already-rotated refresh token"
          }
        },
        "summary": "Refresh the access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "description": "Validates the request and creates the user. When email delivery is configured the account\nis created **unconfirmed** and a confirmation link is emailed — the user must click it\nbefore they can sign in, and no session is issued. When email is not configured the\naccount is activated immediately and a session is issued exactly like [`login`], so\ndevelopment and SMTP-less self-hosting keep working.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Account created; either a confirmation email was sent or (email unconfigured) an access token was issued"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid email, username or password"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Email or username already taken"
          }
        },
        "summary": "Register a new account",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-email": {
      "post": {
        "description": "Consumes a valid, unexpired, unused confirmation token, marks the address verified, and\nsigns the user in — issuing an access token and a rotating refresh cookie exactly like\n[`login`] so clicking the link lands the user in the app.",
        "operationId": "verify_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Email confirmed; session issued"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid or expired confirmation token"
          }
        },
        "summary": "Confirm an email address with a token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-email/resend": {
      "post": {
        "description": "Always responds `202 Accepted`, whether or not the address is registered or already\nconfirmed, so the endpoint can't be used to probe which emails have accounts. A fresh\nlink is only sent when the address exists, is still unconfirmed, and email is configured.",
        "operationId": "resend_verification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "If the address is registered and unconfirmed, a confirmation email has been sent"
          }
        },
        "summary": "Resend the email-confirmation link",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/me/sessions": {
      "get": {
        "description": "The caller's active login sessions (frontend §9.4).",
        "operationId": "sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Active sessions"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "authentication required"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "List active sessions",
        "tags": [
          "me-account"
        ]
      }
    },
...
  },
...
}